+++ commands/dbaas/mariadb/cluster/list.go
@@ -24,7 +24,7 @@ func List() *core.Command {
 		CmdRun: func(c *core.CommandConfig) error {
 			fmt.Fprintf(c.Command.Command.ErrOrStderr(), "%s", jsontabwriter.GenerateVerboseOutput("Getting Clusters..."))
 
-			clusters, err := Clusters(FilterPaginationFlags(c), FilterNameFlags(c))
+			clusters, err := Clusters(FilterNameFlags(c))
 			if err != nil {
 				return err
 			}
@@ -52,8 +52,6 @@ func List() *core.Command {
 	})
 
 	cmd.AddStringFlag(constants.FlagName, constants.FlagNameShort, "", "Response filter to list only the MariaDB Clusters that contain the specified name in the DisplayName field. The value is case insensitive")
-	cmd.AddInt32Flag(constants.FlagMaxResults, constants.FlagMaxResultsShort, 0, constants.DescMaxResults)
-	cmd.AddInt32Flag(constants.FlagOffset, "", 0, "Skip a certain number of results")
 
 	return cmd
 }