+++ commands/dns/record/list.go
@@ -69,8 +69,6 @@ ionosctl dns r list --zone ZONE_ID`,
 		},
 	)
 	cmd.AddStringFlag(constants.FlagName, constants.FlagNameShort, "", "Filter used to fetch only the records that contain specified record name. NOTE: Only available for zone records.")
-	cmd.AddInt32Flag(constants.FlagOffset, "", 0, "The first element (of the total list of elements) to include in the response. Use together with limit for pagination")
-	cmd.AddInt32Flag(constants.FlagMaxResults, "", 0, constants.DescMaxResults)
 
 	cmd.Command.Flags().String(constants.FlagSecondaryZone, "", "The name or ID of the secondary zone to fetch records from")
 	_ = cmd.Command.RegisterFlagCompletionFunc(
@@ -111,12 +109,6 @@ func listRecordsCmd(c *core.CommandConfig) error {
 			}
 			if fn := core.GetFlagName(c.NS, constants.FlagName); viper.IsSet(fn) {
 				req = req.FilterName(viper.GetString(fn))
-			}
-			if fn := core.GetFlagName(c.NS, constants.FlagOffset); viper.IsSet(fn) {
-				req = req.Offset(viper.GetInt32(fn))
-			}
-			if fn := core.GetFlagName(c.NS, constants.FlagMaxResults); viper.IsSet(fn) {
-				req = req.Limit(viper.GetInt32(fn))
 			}
 			return req, nil
 		},
@@ -206,17 +198,6 @@ func secondaryRecords(c *core.CommandConfig) (dns.SecondaryZoneRecordReadList, e
 	}
 
 	req := client.Must().DnsClient.RecordsApi.SecondaryzonesRecordsGet(context.Background(), secondaryZoneID)
-
-	if c.Command.Command.Flags().Changed(constants.FlagOffset) {
-		offset, _ := c.Command.Command.Flags().GetInt32(constants.FlagOffset)
-		req = req.Offset(offset)
-	}
-
-	if c.Command.Command.Flags().Changed(constants.FlagMaxResults) {
-		maxResults, _ := c.Command.Command.Flags().GetInt32(constants.FlagMaxResults)
-		req = req.Limit(maxResults)
-	}
-
 	records, _, err := req.Execute()
 	if err != nil {
 		return dns.SecondaryZoneRecordReadList{}, err