+++ commands/dns/reverse-record/list.go
@@ -9,7 +9,7 @@ import (
 	"github.com/ionos-cloud/ionosctl/v6/internal/printer/json2table/jsonpaths"
 	"github.com/ionos-cloud/ionosctl/v6/internal/printer/jsontabwriter"
 	"github.com/ionos-cloud/ionosctl/v6/internal/printer/tabheaders"
-	dns "github.com/ionos-cloud/sdk-go-bundle/products/dns/v2"
+	"github.com/ionos-cloud/sdk-go-bundle/products/dns/v2"
 )
 
 func List() *core.Command {
@@ -21,7 +21,7 @@ func List() *core.Command {
 		ShortDesc: "Retrieve all reverse records",
 		Example:   "ionosctl dns rr list",
 		CmdRun: func(c *core.CommandConfig) error {
-			ls, err := Records(FilterLimitOffset(c.NS), FilterRecordsByIp(c.NS))
+			ls, err := Records(FilterRecordsByIp(c.NS))
 			if err != nil {
 				return fmt.Errorf("failed listing records: %w", err)
 			}
@@ -51,8 +51,6 @@ func List() *core.Command {
 			})
 		}, constants.DNSApiRegionalURL, constants.DNSLocations),
 	)
-	cmd.AddInt32Flag(constants.FlagOffset, "", 0, "The first element (of the total list of elements) to include in the response. Use together with limit for pagination")
-	cmd.AddInt32Flag(constants.FlagMaxResults, "", 0, constants.DescMaxResults)
 
 	return cmd
 }