+++ internal/client/builder.go
@@ -1,83 +1,32 @@
 package client
 
 import (
-	"fmt"
 	"net/url"
 
-	"github.com/ionos-cloud/sdk-go-bundle/products/apigateway/v2"
-	"github.com/ionos-cloud/sdk-go-bundle/products/dbaas/inmemorydb/v2"
-	"github.com/ionos-cloud/sdk-go-bundle/products/monitoring/v2"
-	"github.com/ionos-cloud/sdk-go-bundle/shared"
-	"github.com/ionos-cloud/sdk-go-bundle/shared/fileconfiguration"
-
 	"github.com/ionos-cloud/ionosctl/v6/internal/constants"
+	"github.com/ionos-cloud/sdk-go-bundle/products/apigateway/v2"
 	"github.com/ionos-cloud/sdk-go-bundle/products/auth/v2"
 	"github.com/ionos-cloud/sdk-go-bundle/products/cdn/v2"
 	"github.com/ionos-cloud/sdk-go-bundle/products/cert/v2"
 	"github.com/ionos-cloud/sdk-go-bundle/products/containerregistry/v2"
+	"github.com/ionos-cloud/sdk-go-bundle/products/dbaas/inmemorydb/v2"
 	"github.com/ionos-cloud/sdk-go-bundle/products/dbaas/mariadb/v2"
 	"github.com/ionos-cloud/sdk-go-bundle/products/dbaas/mongo/v2"
 	"github.com/ionos-cloud/sdk-go-bundle/products/dbaas/psql/v2"
 	"github.com/ionos-cloud/sdk-go-bundle/products/dns/v2"
 	"github.com/ionos-cloud/sdk-go-bundle/products/kafka/v2"
 	"github.com/ionos-cloud/sdk-go-bundle/products/logging/v2"
+	"github.com/ionos-cloud/sdk-go-bundle/products/monitoring/v2"
 	"github.com/ionos-cloud/sdk-go-bundle/products/vpn/v2"
+	"github.com/ionos-cloud/sdk-go-bundle/shared"
 	vmasc "github.com/ionos-cloud/sdk-go-vm-autoscaling"
 	cloudv6 "github.com/ionos-cloud/sdk-go/v6"
 
 	"github.com/spf13/viper"
 )
 
-// AuthSource represents a human-readable description of where the client's authentication credentials were sourced from.
-type AuthSource string
-
-const (
-	AuthSourceEnvBearer AuthSource = "environment variable: IONOS_TOKEN"
-	AuthSourceEnvBasic  AuthSource = "environment variables: IONOS_USERNAME, IONOS_PASSWORD"
-	AuthSourceCfgBearer AuthSource = "credentials from config file: token"
-	AuthSourceCfgBasic  AuthSource = "credentials from config file: username, password"
-	AuthSourceNone      AuthSource = "no authentication provided"
-)
-
-// all possible sources in priority order
-var AuthOrder = []AuthSource{
-	AuthSourceEnvBearer,
-	AuthSourceEnvBasic,
-	AuthSourceCfgBearer,
-	AuthSourceCfgBasic,
-}
-
-type Client struct {
-	Config      *fileconfiguration.FileConfig
-	ConfigPath  string // Path to the config file used to create this client, if any.
-	AuthSource  AuthSource
-	URLOverride string // If the client was created with a specific URL override, this will hold that value. If we notice a change in the URL, we need to re-create the client.
-
-	Apigateway           *apigateway.APIClient
-	CloudClient          *cloudv6.APIClient
-	AuthClient           *auth.APIClient
-	CertManagerClient    *cert.APIClient
-	RegistryClient       *containerregistry.APIClient
-	DnsClient            *dns.APIClient
-	LoggingServiceClient *logging.APIClient
-	VMAscClient          *vmasc.AutoScalingGroupsApiService
-	VPNClient            *vpn.APIClient
-	CDNClient            *cdn.APIClient
-	Kafka                *kafka.APIClient
-	Monitoring           *monitoring.APIClient
-
-	PostgresClient   *psql.APIClient
-	MongoClient      *mongo.APIClient
-	MariaClient      *mariadb.APIClient
-	InMemoryDBClient *inmemorydb.APIClient
-}
-
-func appendUserAgent(userAgent string) string {
-	return fmt.Sprintf("%v_%v", viper.GetString(constants.CLIHttpUserAgent), userAgent)
-}
-
 // hostWithoutPath strips any path from hostUrl; so that SDK clients append their own product paths,
-// avoiding double basepaths ('/databases/postgresql/cloudapi/v6')
+// thus avoiding double basepaths ('/databases/postgresql/cloudapi/v6')
 // If for some reason this needs to be removed in the future, then please remove
 // the default basepaths in all 'WithConfigOverride' calls too.
 func hostWithoutPath(h string) string {
@@ -129,6 +78,15 @@ func newClient(name, pwd, token, hostUrl string) *Client {
 		// don't explicitly set to Off, as this breaks SDK handling of the IONOS_LOG_LEVEL variable
 	}
 
+	queryParams := map[string]string{
+		"limit":  viper.GetString(constants.FlagLimit),
+		"offset": viper.GetString(constants.FlagOffset),
+	}
+
+	setQueryParams(sharedConfig, queryParams)
+	setQueryParams(clientConfig, queryParams)
+	setQueryParams(vmascConfig, queryParams)
+
 	return &Client{
 		URLOverride: hostUrl,
 
@@ -153,5 +111,15 @@ func newClient(name, pwd, token, hostUrl string) *Client {
 
 		MariaClient:      mariadb.NewAPIClient(sharedConfig),
 		InMemoryDBClient: inmemorydb.NewAPIClient(sharedConfig),
+	}
+}
+
+type hasQueryParam interface {
+	AddDefaultQueryParam(key, val string)
+}
+
+func setQueryParams(cfg hasQueryParam, params map[string]string) {
+	for k, v := range params {
+		cfg.AddDefaultQueryParam(k, v)
 	}
 }