+++ services/cloudapi-v6/resources/datacenter.go
@@ -72,9 +72,6 @@ func (ds *dataCentersService) List(params ListQueryParams) (Datacenters, *Respon
 		if params.OrderBy != nil {
 			req = req.OrderBy(*params.OrderBy)
 		}
-		if params.MaxResults != nil {
-			req = req.MaxResults(*params.MaxResults)
-		}
 		if !structs.IsZero(params.QueryParams) {
 			if params.QueryParams.Depth != nil {
 				req = req.Depth(*params.QueryParams.Depth)