+++ services/cloudapi-v6/resources/firewallrule.go
@@ -57,9 +57,6 @@ func (svc *firewallRulesService) List(datacenterId, serverId, nicId string, para
 		if params.OrderBy != nil {
 			req = req.OrderBy(*params.OrderBy)
 		}
-		if params.MaxResults != nil {
-			req = req.MaxResults(*params.MaxResults)
-		}
 		if !structs.IsZero(params.QueryParams) {
 			if params.QueryParams.Depth != nil {
 				req = req.Depth(*params.QueryParams.Depth)