+++ services/cloudapi-v6/resources/flowlog.go
@@ -61,9 +61,6 @@ func (svc *flowLogsService) List(datacenterId, serverId, nicId string, params Li
 		if params.OrderBy != nil {
 			req = req.OrderBy(*params.OrderBy)
 		}
-		if params.MaxResults != nil {
-			req = req.MaxResults(*params.MaxResults)
-		}
 		if !structs.IsZero(params.QueryParams) {
 			if params.QueryParams.Depth != nil {
 				req = req.Depth(*params.QueryParams.Depth)