+++ services/cloudapi-v6/resources/location.go
@@ -55,9 +55,6 @@ func (s *locationsService) List(params ListQueryParams) (Locations, *Response, e
 		if params.OrderBy != nil {
 			req = req.OrderBy(*params.OrderBy)
 		}
-		if params.MaxResults != nil {
-			req = req.MaxResults(*params.MaxResults)
-		}
 		if !structs.IsZero(params.QueryParams) {
 			if params.QueryParams.Depth != nil {
 				req = req.Depth(*params.QueryParams.Depth)