+++ services/cloudapi-v6/resources/query_params.go
@@ -3,7 +3,6 @@ package resources
 type ListQueryParams struct {
 	Filters     *map[string][]string `json:"Filters,omitempty"`
 	OrderBy     *string              `json:"OrderBy,omitempty"`
-	MaxResults  *int32               `json:"MaxResults,omitempty"`
 	QueryParams QueryParams          `json:"QueryParams,omitempty"`
 }
 
@@ -19,13 +18,6 @@ func (q ListQueryParams) SetOrderBy(orderBy string) ListQueryParams {
 	return q
 }
 
-func (q ListQueryParams) SetMaxResults(maxResults int32) ListQueryParams {
-	if maxResults > 0 {
-		q.MaxResults = &maxResults
-	}
-	return q
-}
-
 func (q ListQueryParams) SetDepth(depth int32) ListQueryParams {
 	q.QueryParams.Depth = &depth
 	return q
@@ -35,9 +27,6 @@ func (q ListQueryParams) SetPretty(pretty bool) ListQueryParams {
 	q.QueryParams.Pretty = &pretty
 	return q
 }
-
-// TODO: Merge ListQueryParams into QueryParams
-// TODO: Once Compute namespace is added, add all of QueryParam's flags as Global flags, to reduce duplication
 
 type QueryParams struct {
 	Depth  *int32 `json:"Depth,omitempty"`