+++ services/cloudapi-v6/resources/request.go
@@ -56,9 +56,6 @@ func (rs *requestsService) List(params ListQueryParams) (Requests, *Response, er
 		if params.OrderBy != nil {
 			req = req.OrderBy(*params.OrderBy)
 		}
-		if params.MaxResults != nil {
-			req = req.MaxResults(*params.MaxResults)
-		}
 		if !structs.IsZero(params.QueryParams) {
 			if params.QueryParams.Depth != nil {
 				req = req.Depth(*params.QueryParams.Depth)