+++ services/cloudapi-v6/resources/snapshot.go
@@ -58,9 +58,6 @@ func (s *snapshotsService) List(params ListQueryParams) (Snapshots, *Response, e
 		if params.OrderBy != nil {
 			req = req.OrderBy(*params.OrderBy)
 		}
-		if params.MaxResults != nil {
-			req = req.MaxResults(*params.MaxResults)
-		}
 		if !structs.IsZero(params.QueryParams) {
 			if params.QueryParams.Depth != nil {
 				req = req.Depth(*params.QueryParams.Depth)