+++ services/cloudapi-v6/resources/user.go
@@ -84,9 +84,6 @@ func (s *usersService) List(params ListQueryParams) (Users, *Response, error) {
 		if params.OrderBy != nil {
 			req = req.OrderBy(*params.OrderBy)
 		}
-		if params.MaxResults != nil {
-			req = req.MaxResults(*params.MaxResults)
-		}
 		if !structs.IsZero(params.QueryParams) {
 			if params.QueryParams.Depth != nil {
 				req = req.Depth(*params.QueryParams.Depth)